Accidentally a ceilf() crept in.
(Note; sorting this out once, for all platforms, could be worth the work.
float math can be faster in most cpus, no?)


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@4356 954f8c5b-7b00-dc11-b283-0030488c597c

+++ blender/source/blender/blenkernel/intern/softbody.c
@@ -1097,7 +1097,7 @@ static void softbody_baked_add(Object *ob, float framenr)
 		if(sb->sfra >= sb->efra) return;		// safety, UI or py setting allows
 		if(sb->interval<1) sb->interval= 1;		// just be sure
 		
-		sb->totkey= 1 + (int)(ceilf( (efra-sfra)/dfra ) );
+		sb->totkey= 1 + (int)(ceil( (efra-sfra)/dfra ) );
 		sb->keys= MEM_callocN( sizeof(void *)*sb->totkey, "sb keys");
 	}
